+++ dev/App.vue
@@ -78,6 +78,23 @@
           </ax-form-switch>
         </ax-form-field>
       </ax-form>
+
+      <div class="mt-5"></div>
+
+      <ax-btn class="airforce dark-1 rounded-1 shadow-1" @click="isModalOpened = !isModalOpened"
+        >Toggle modal</ax-btn
+      >
+
+      <ax-modal class="white rounded-1 shadow-1" v-model="isModalOpened">
+        Hey guys
+
+        <template v-slot:footer>
+          <ax-btn class="airforce dark-1 rounded-1 shadow-1" @click="isModal2Opened = !isModal2Opened"
+            >Toggle modal 2</ax-btn
+          >
+        </template>
+      </ax-modal>
+      <ax-modal class="red rounded-1 shadow-1" v-model="isModal2Opened"> HOLA </ax-modal>
     </main>
   </div>
 </template>
@@ -91,6 +108,8 @@ export default {
     isCollapsibleOpened2: true,
     radio: 'Yes',
     checked: true,
+    isModalOpened: false,
+    isModal2Opened: false,
   }),
   watch: {
     radio(val) {
@@ -98,6 +117,9 @@ export default {
     },
     checked(state) {
       console.log('check', state);
+    },
+    isModalOpened(state) {
+      console.log('modalOpened', state);
     },
   },
 };