+++ src/components/index.js
@@ -1,6 +1,7 @@
 import AxBtn from './button';
 import AxSidenav from './sidenav';
 import AxForms from './forms';
+import AxModal from './modal';
 import AxCollapsible from './collapsible';
 
-export { AxBtn, AxSidenav, AxForms, AxCollapsible };
+export { AxBtn, AxSidenav, AxForms, AxModal, AxCollapsible };